Fix the bounce buffer size in ORC writer (#14947)
Closes #14932

ORC writer uses uncompressed stream sizes when allocating the bounce buffer. This can lead to issues when all uncompressed streams are larger than the GDS threshold, but compressed size is not. In this scenario, the bounce buffer is not allocated, and writing the compressed stream through the bounce buffer causes a crash.

This PR moves the computation of the bounce buffer size until after compression, so it works with correct stream sizes.

Authors:
   - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
   - Nghia Truong (https://github.com/ttnghia)
   - Bradley Dice (https://github.com/bdice)