Benchmark decimal <--> floating conversions. (#15334)
Adds benchmarks for decimal <--> floating conversions.  Does so for float <--> decimal32 & decimal64, and for double <--> decimal32, decimal64, and decimal128.  Within a column data tends to be in a similar range of values, so this provides separate tests for different representative ranges of powers-of-10.  

Note that with the current conversion algorithm, the max supported scale of a decimal is the max power of 10 that that type can hold, so scale 9 for decimal32, 19 for decimal64, and 38 for decimal128. Thus only these ranges of floats/doubles are tested. 

Also adds the ability to generate decimals with a specific (rather than random) scale factor. This expands the API, it does not replace the existing one.  All existing tests that generate a column of random decimals will continue to do so with a random scale factor, this capability is opt-in.  The machinery for this was already there, but only partially; this change fills it in.

Authors:
  - Paul Mattione (https://github.com/pmattione-nvidia)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cudf/pull/15334