Add patch for incorrect cuco noexcept clauses (#16077)
[cuco previously marked a number of methods as noexcept that can in fact
throw exceptions](https://github.com/nvidia/cuCollections/issues/510).
This causes problems for cudf functions that call these methods. The
issue [was fixed in cuco
upstream](https://github.com/NVIDIA/cuCollections/pull/511), but we
cannot easily update to the latest commit of cuco, especially in a patch
fix for 24.06. This PR instead adds a rapids-cmake patch for the cuco
clone to address this issue. The patch may be removed once we update to
a commit of cuco that contains the necessary fix.

Resolves #16059